Feat/re sync missing order

+++ internal/model/aff_order.go
@@ -40,9 +40,9 @@ type AffOrder struct {
 	LandingPage        string     `json:"landing_page"`
 	Merchant           string     `json:"merchant"`
 	AccessTradeOrderId string     `json:"accesstrade_order_id" gorm:"column:accesstrade_order_id"`
-	OrderPending       *uint8     `json:"order_pending"`
-	OrderReject        *uint8     `json:"order_reject"`
-	OrderApproved      *uint8     `json:"order_approved"`
+	OrderPending       *uint      `json:"order_pending"`
+	OrderReject        *uint      `json:"order_reject"`
+	OrderApproved      *uint      `json:"order_approved"`
 	ProductCategory    string     `json:"product_category"`
 	ProductsCount      int        `json:"products_count"`
 	PubCommission      float32    `json:"pub_commission"`
